chore: Remove dead code (unused subsystems and stub files)
Removes code that is unreachable from any entry point, verified with
`deadcode` and `staticcheck`:

- pkg/util/net.go: GetFirstAddressOf + filterIPs. Superseded by dskit's
  netutil.GetFirstAddressOf, which is what the codebase actually calls.
- pkg/util/events.go: the Event/InitEvents sampling logger. Its own
  comment called it a "temporary hack implementation"; it has no callers.
- pkg/querier/astmapper: the shardSummer ASTMapper (NewShardSummer and
  helpers), a Prometheus-PromQL sharding mapper with no callers. Loki
  shards via its own LogQL path. The live Shard* label primitives are
  kept, and since that's all the file holds now, it's renamed
  shard_summer.go -> shard_annotation.go.
- Empty/placeholder stub files in logcli/output, indexshipper/util,
  engine/internal/types, and ruler/rulestore/objectclient.
