--- pkg/engine/internal/types/util.go
@@ -1,1 +1,0 @@
-package types //nolint:revive
