--- pkg/logcli/output/loki.go
@@ -1,1 +1,0 @@
-package output
