--- pkg/ruler/rulestore/objectclient/rule_store_test.go
@@ -1,4 +1,0 @@
-package objectclient
-
-// This is an empty placeholder test file to signal that the tests for RuleStore are in:
-// pkg/ruler/rulestore/bucketclient/bucket_client_test.go
