--- pkg/storage/stores/shipper/indexshipper/util/util.go
@@ -1,1 +1,0 @@
-package util //nolint:revive
